First working version of the ROS API checker.  This tool verifies the presences of a ROS API that is described in a YAML file.  It'll take in multiple YAML files, merging their contents, which makes it easy to break the API definition up into reusable pieces.  For each topic given in the API definition, the tool will verify the presence of the topic, its type, and the number of publishers and subscribers (-1 means don't care for the publisher/subscriber count).
In this version, there's a test that verifies the API provided by
atlas.launch (which is Atlas + Multisense-SL) and another for
atlas_sandia_hands.launch (Atlas + Multisense-SL + Sandia hands).  New tests can
be added or composed in a similar fashion.

The tool supports a `strict` flag; if set, it will perform an extra test to
compare the output of `rostopic list` to the set of topics given in the API
definition.  If they're not equal (which means that some extra, unexpected
topics, are either advertised or subscribed), that test will fail.

The `strict` flag is not set on the current tests because there's some API
cleanup yet to be done.

This tool should eventually migrate out of drcsim and into ROS somewhere, likely
rostest.