DO NOT MERGE FROM HERE.
If you like these changes, then approve pull request #173

Manually applied the diff between AtlasSimInterface_1.0.5 and udp_tweak,
excluding changes in the plugins/ directory.  The goal is produce a pull request
that bitbucket can render for review in place of #173, which has huge changes in
the BDI library.

Specifically, I did the following:
    hg diff -r udp_tweak -r AtlasSimInterface_1.0.5 --exclude "plugins/*" > /tmp/d
    hg up udp_tweak
    hg branch surrogate_for_173_do_not_merge
    patch  -p1 < /tmp/d